requirements.txt
django==4.2.2
djangorestframework
djangorestframework-simplejwt